refactor: extract types in their own file
