+++ src/libs/recommendation/useHomeRecommendedIdsMutation.ts
@@ -3,29 +3,7 @@ import { useMutation } from 'react-query'
 import { analytics } from 'libs/firebase/analytics'
 import { eventMonitoring } from 'libs/monitoring'
 import { QueryKeys } from 'libs/queryKeys'
-
-export interface RecommendedIdsResponse {
-  playlist_recommended_offers: string[]
-  params: {
-    reco_origin?: string
-    model_name?: string
-    model_version?: string
-    geo_located?: boolean
-    ab_test?: string
-    filtered?: boolean
-  }
-}
-
-export interface RecommendedIdsRequest {
-  start_date?: string
-  end_date?: string
-  isEvent?: boolean
-  categories?: string[]
-  price_max?: number
-  subcategories?: string[]
-  isDuo?: boolean
-  isRecoShuffled?: boolean
-}
+import { RecommendedIdsRequest, RecommendedIdsResponse } from 'libs/recommendation/types'
 
 export const useHomeRecommendedIdsMutation = (recommendationUrl: string) => {
   return useMutation(
